Daily AlpacaHack/Stateless Auth/attachment/web/requirements.txt
Flask==3.1.2
PyJWT==2.10.1